feat: unify start time component and add driver control nav
