+++ pages/upload.tsx
@@ -2,6 +2,7 @@ import { useState, useEffect, DragEvent, ChangeEvent } from 'react';
 import Layout from '../components/Layout';
 import Icon from '../components/Icon';
 import UploadHistory from '../components/UploadHistory';
+import DCNavbar from '../components/DCNavbar';
 
 export default function Upload() {
   const [files, setFiles] = useState<File[]>([]);
@@ -117,7 +118,8 @@ export default function Upload() {
   };
 
   return (
-    <Layout title="Upload Files">
+    <Layout title="Upload Files" hideNavbar fullWidth>
+      <DCNavbar />
       <h1 className="text-2xl font-bold mb-4">Upload Files</h1>
       <div
         className="border-2 border-dashed rounded-lg p-6 text-center bg-white dark:bg-gray-700 dark:border-gray-500 dark:text-gray-200 cursor-pointer"
